Ensure even video dimensions (#82)
gotcha, I think we may have to do some work iOS later to properly grab the actual capture dimensions rather than rely on the dimensions we request, to properly do the calculations.

That said, I think m97 is now handling the odd dimensions properly, so I don't think we'll need this fix anymore.

As for adaptOutputFormat, flutter_webrtc is already doing that (though it doesn't seem to prevent the crash on m93). Should native android use this as well?